requirements.txt
numpy==1.24.4
pandas==1.5.3
matplotlib==3.7.2
seaborn==0.12.2
transformers==4.42.4
torch==2.3.1+cu121
Requests==2.32.3
nltk==3.6.3
streamlit==1.33.0
python-dotenv
